requirements.txt
Django==2.2.1
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
